backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.0.0
